Update packages.qmd - fishboot package added to the list.
The fishboot package contains a suite of new bootstrap-based models and tools for the study of fish stocks and aquatic populations. It is intended for ecologists and fisheries scientists who study the body growth and mortality of fish and macroinvertebrates. Data sources for fishboot functions are length-frequency distributions, tag-and-recapture, and readings of annual rings in hard structures (e.g., otoliths). See Schwamborn et al., 2019 <doi:10.1016/j.ecolmodel.2018.12.001> for an overview. Includes bootstrapped curve fit functions and several plotting functions. 

Best Regards,

Ralf Schwamborn